requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
tokenizers>=0.14.0

safetensors>=0.4.0
huggingface-hub>=0.19.0

chromadb>=0.4.22
sentence-transformers>=2.2.0

PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0
PyQt6-sip>=13.6.0

SQLAlchemy>=2.0.0
aiosqlite>=0.19.0

PyYAML>=6.0.1
python-dotenv>=1.0.0
rich>=13.7.0
tqdm>=4.66.0

numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0

datasets>=2.16.0
requests>=2.31.0
beautifulsoup4>=4.12.0
psutil>=5.9.0
pytest>=7.4.0
pytest-qt>=4.2.0
coverage>=7.4.0
